+++ src/api/sitzung.rs
@@ -1,20 +1,19 @@
+use crate::db::retrieve::{SitzungFilterParameters, sitzung_by_param};
 use crate::db::{delete, insert, retrieve};
 use crate::error::LTZFError;
 use crate::{LTZFServer, Result};
 use async_trait::async_trait;
 use axum::http::Method;
 use axum_extra::extract::{CookieJar, Host};
+use chrono::Datelike;
 use openapi::apis::adminschnittstellen_collector_schnittstellen_kalender_sitzungen::*;
 use openapi::apis::adminschnittstellen_sitzungen::*;
 use openapi::apis::kalender_sitzungen_unauthorisiert::*;
 use openapi::apis::sitzungen_unauthorisiert::*;
 use openapi::models;
-use openapi::models::SGetQueryParams;
-use openapi::models::*;
-use sqlx::PgTransaction;
 
 use super::auth::{self, APIScope};
-use super::{PaginationResponsePart, compare::*};
+use super::{PaginationResponsePart, compare::*, find_applicable_date_range};
 
 #[async_trait]
 impl AdminschnittstellenSitzungen<LTZFError> for LTZFServer {
@@ -101,40 +100,43 @@ impl KalenderSitzungenUnauthorisiert<LTZFError> for LTZFServer {
     #[allow(clippy::type_complexity, clippy::type_repetition_in_bounds)]
     async fn kal_date_get(
         &self,
-        method: &Method,
-        host: &Host,
-        cookies: &CookieJar,
+        _method: &Method,
+        _host: &Host,
+        _cookies: &CookieJar,
         header_params: &models::KalDateGetHeaderParams,
         path_params: &models::KalDateGetPathParams,
         query_params: &models::KalDateGetQueryParams,
     ) -> Result<KalDateGetResponse> {
         let mut tx = self.sqlx_db.begin().await?;
-        let date = path_params.datum;
-        let dt_begin = date
-            .and_time(chrono::NaiveTime::from_hms_micro_opt(0, 0, 0, 0).unwrap())
-            .and_utc();
-        let dt_end = date
-            .checked_add_days(chrono::Days::new(1))
-            .unwrap()
-            .and_time(chrono::NaiveTime::from_hms_micro_opt(0, 0, 0, 0).unwrap())
-            .and_utc();
-
-        // header building
-        let x_total_count = sqlx::query!(
-            "SELECT count(1) as c FROM sitzung s
-        INNER JOIN gremium g ON g.id = s.gr_id
-        INNER JOIN parlament p ON p.id = g.parl 
-        WHERE termin BETWEEN $1 AND $2 AND p.value = $3",
-            dt_begin,
-            dt_end,
-            path_params.parlament.to_string()
+        let dr = find_applicable_date_range(
+            Some(path_params.datum.year() as u32),
+            Some(path_params.datum.month()),
+            Some(path_params.datum.day()),
+            None,
+            None,
+            header_params.if_modified_since,
         )
-        .map(|r| r.c)
-        .fetch_one(&mut *tx)
-        .await?
-        .map(|x| x as i32);
+        .expect("Path parameter datum is mandatory");
+
+        let dt_begin = dr.since;
+        let dt_end = dr.until;
+        let values = sitzung_by_param(
+            &SitzungFilterParameters {
+                parlament: Some(path_params.parlament),
+                gremium_like: None,
+                since: dt_begin,
+                until: dt_end,
+                vgid: None,
+                wp: None,
+            },
+            query_params.page,
+            query_params.per_page,
+            &mut tx,
+        )
+        .await?;
+
         let prp = PaginationResponsePart::new(
-            x_total_count,
+            Some(values.0 as i32),
             query_params.page,
             query_params.per_page,
             &format!(
@@ -144,37 +146,17 @@ impl KalenderSitzungenUnauthorisiert<LTZFError> for LTZFServer {
             ),
         );
 
-        // actual fetch
-        let sids = sqlx::query!(
-            "SELECT s.id FROM sitzung s 
-        INNER JOIN gremium g ON g.id = s.gr_id
-        INNER JOIN parlament p ON p.id = g.parl 
-        WHERE termin BETWEEN $1 AND $2 AND p.value = $3
-        LIMIT $4 OFFSET $5",
-            dt_begin,
-            dt_end,
-            path_params.parlament.to_string(),
-            prp.limit(),
-            prp.offset()
-        )
-        .map(|r| r.id)
-        .fetch_all(&mut *tx)
-        .await?;
-        if sids.is_empty() {
+        if values.1.is_empty() {
+            tx.rollback().await?;
             return Ok(KalDateGetResponse::Status404_NotFound {
                 x_rate_limit_limit: None,
                 x_rate_limit_remaining: None,
                 x_rate_limit_reset: None,
             });
         }
-        let mut vector = vec![];
-        for sid in sids {
-            vector.push(retrieve::sitzung_by_id(sid, &mut tx).await?);
-        }
-
         tx.commit().await?;
         Ok(KalDateGetResponse::Status200_SuccessfulResponse {
-            body: vector,
+            body: values.1,
             x_rate_limit_limit: None,
             x_rate_limit_remaining: None,
             x_rate_limit_reset: None,
@@ -231,9 +213,21 @@ impl KalenderSitzungenUnauthorisiert<LTZFError> for LTZFServer {
         let result =
             retrieve::sitzung_by_param(&params, query_params.page, query_params.per_page, &mut tx)
                 .await?;
-        if result.is_empty() {
+        let prp = PaginationResponsePart::new(
+            Some(result.0 as i32),
+            query_params.page,
+            query_params.per_page,
+            "/api/v1/kalender",
+        );
+        if result.1.is_empty() {
             tx.rollback().await?;
             Ok(KalGetResponse::Status204_NoContent {
+                x_rate_limit_limit: None,
+                x_rate_limit_remaining: None,
+                x_rate_limit_reset: None,
+            })
+        } else if result.1.is_empty() && header_params.if_modified_since.is_some() {
+            Ok(KalGetResponse::Status304_NotModified {
                 x_rate_limit_limit: None,
                 x_rate_limit_remaining: None,
                 x_rate_limit_reset: None,
@@ -241,15 +235,15 @@ impl KalenderSitzungenUnauthorisiert<LTZFError> for LTZFServer {
         } else {
             tx.commit().await?;
             Ok(KalGetResponse::Status200_SuccessfulResponse {
-                body: result,
+                body: result.1,
                 x_rate_limit_limit: None,
                 x_rate_limit_remaining: None,
                 x_rate_limit_reset: None,
-                x_total_count: (),
-                x_total_pages: (),
-                x_page: (),
-                x_per_page: (),
-                link: (),
+                x_total_count: prp.x_total_count,
+                x_total_pages: prp.x_total_pages,
+                x_page: prp.x_page,
+                x_per_page: prp.x_per_page,
+                link: prp.link,
             })
         }
     }
@@ -320,9 +314,9 @@ impl SitzungenUnauthorisiert<LTZFError> for LTZFServer {
     #[allow(clippy::type_complexity, clippy::type_repetition_in_bounds)]
     async fn s_get(
         &self,
-        method: &Method,
-        host: &Host,
-        cookies: &CookieJar,
+        _method: &Method,
+        _host: &Host,
+        _cookies: &CookieJar,
         header_params: &models::SGetHeaderParams,
         query_params: &models::SGetQueryParams,
     ) -> Result<SGetResponse> {
@@ -343,8 +337,6 @@ impl SitzungenUnauthorisiert<LTZFError> for LTZFServer {
         }
         let params = retrieve::SitzungFilterParameters {
             gremium_like: None,
-            limit: query_params.limit.map(|x| x as u32),
-            offset: query_params.offset.map(|x| x as u32),
             parlament: query_params.p,
             wp: query_params.wp.map(|x| x as u32),
             since: range.as_ref().unwrap().since,
@@ -353,15 +345,23 @@ impl SitzungenUnauthorisiert<LTZFError> for LTZFServer {
         };
 
         let mut tx: sqlx::Transaction<'_, sqlx::Postgres> = self.sqlx_db.begin().await?;
-        let result = retrieve::sitzung_by_param(&params, &mut tx).await?;
+        let result =
+            retrieve::sitzung_by_param(&params, query_params.page, query_params.per_page, &mut tx)
+                .await?;
+        let prp = PaginationResponsePart::new(
+            Some(result.0 as i32),
+            query_params.page,
+            query_params.per_page,
+            "/api/v1/sitzung",
+        );
         tx.commit().await?;
-        if result.is_empty() && header_params.if_modified_since.is_none() {
+        if result.1.is_empty() && header_params.if_modified_since.is_none() {
             Ok(SGetResponse::Status204_NoContent {
                 x_rate_limit_limit: None,
                 x_rate_limit_remaining: None,
                 x_rate_limit_reset: None,
             })
-        } else if result.is_empty() && header_params.if_modified_since.is_some() {
+        } else if result.1.is_empty() && header_params.if_modified_since.is_some() {
             Ok(SGetResponse::Status304_NotModified {
                 x_rate_limit_limit: None,
                 x_rate_limit_remaining: None,
@@ -369,15 +369,15 @@ impl SitzungenUnauthorisiert<LTZFError> for LTZFServer {
             })
         } else {
             Ok(SGetResponse::Status200_SuccessfulResponse {
-                body: result,
+                body: result.1,
                 x_rate_limit_limit: None,
                 x_rate_limit_remaining: None,
                 x_rate_limit_reset: None,
-                x_total_count: (),
-                x_total_pages: (),
-                x_page: (),
-                x_per_page: (),
-                link: (),
+                x_total_count: prp.x_total_count,
+                x_total_pages: prp.x_total_pages,
+                x_page: prp.x_page,
+                x_per_page: prp.x_per_page,
+                link: prp.link,
             })
         }
     }
@@ -469,308 +469,6 @@ impl AdminschnittstellenCollectorSchnittstellenKalenderSitzungen<LTZFError> for 
             x_rate_limit_remaining: None,
             x_rate_limit_reset: None,
         })
-    }
-}
-
-pub struct DateRange {
-    pub since: Option<chrono::DateTime<chrono::Utc>>,
-    pub until: Option<chrono::DateTime<chrono::Utc>>,
-}
-impl
-    From<(
-        Option<chrono::DateTime<chrono::Utc>>,
-        Option<chrono::DateTime<chrono::Utc>>,
-    )> for DateRange
-{
-    fn from(
-        value: (
-            Option<chrono::DateTime<chrono::Utc>>,
-            Option<chrono::DateTime<chrono::Utc>>,
-        ),
-    ) -> Self {
-        Self {
-            since: value.0,
-            until: value.1,
-        }
-    }
-}
-pub fn find_applicable_date_range(
-    y: Option<u32>,
-    m: Option<u32>,
-    d: Option<u32>,
-    since: Option<chrono::DateTime<chrono::Utc>>,
-    until: Option<chrono::DateTime<chrono::Utc>>,
-    ifmodsince: Option<chrono::DateTime<chrono::Utc>>,
-) -> Option<DateRange> {
-    let ymd_date_range = if let Some(y) = y {
-        if let Some(m) = m {
-            if let Some(d) = d {
-                Some((
-                    chrono::NaiveDate::from_ymd_opt(y as i32, m, d).unwrap(),
-                    chrono::NaiveDate::from_ymd_opt(y as i32, m, d).unwrap(),
-                ))
-            } else {
-                Some((
-                    chrono::NaiveDate::from_ymd_opt(y as i32, m, 1).unwrap(),
-                    chrono::NaiveDate::from_ymd_opt(y as i32, m + 1, 1)
-                        .unwrap()
-                        .checked_sub_days(chrono::Days::new(1))
-                        .unwrap(),
-                ))
-            }
-        } else {
-            Some((
-                chrono::NaiveDate::from_ymd_opt(y as i32, 1, 1).unwrap(),
-                chrono::NaiveDate::from_ymd_opt(y as i32, 12, 31).unwrap(),
-            ))
-        }
-    } else {
-        None
-    }
-    .map(|(a, b)| {
-        (
-            a.and_hms_opt(0, 0, 0).unwrap().and_utc(),
-            b.and_hms_opt(23, 59, 59).unwrap().and_utc(),
-        )
-    });
-
-    let mut since_min = ifmodsince;
-    let mut until_min = until;
-    if since.is_some() {
-        if since_min.is_some() {
-            since_min = Some(since_min.unwrap().min(since.unwrap()));
-        } else {
-            since_min = since;
-        }
-    }
-    if let Some((ymd_s, ymd_u)) = ymd_date_range {
-        if since_min.is_some() {
-            since_min = Some(ymd_s.max(since_min.unwrap()));
-        } else {
-            since_min = Some(ymd_s);
-        }
-        if until_min.is_some() {
-            until_min = Some(ymd_u.min(until_min.unwrap()));
-        } else {
-            until_min = Some(ymd_u);
-        }
-    }
-
-    // semantic check
-    if let Some(sm) = since_min {
-        if sm < chrono::DateTime::parse_from_rfc3339("1945-01-01T00:00:00+00:00").unwrap() {
-            return None;
-        }
-        if let Some(um) = until {
-            if sm >= um {
-                return None;
-            }
-        }
-    }
-
-    if let Some((ys, yu)) = ymd_date_range {
-        if since_min.is_some() && since_min.unwrap() > yu
-            || until_min.is_some() && until_min.unwrap() < ys
-        {
-            None
-        } else {
-            Some((since_min, until_min).into())
-        }
-    } else {
-        Some((since_min, until_min).into())
-    }
-}
-
-pub async fn kal_get_by_param(
-    qparams: &KalGetQueryParams,
-    hparams: &KalGetHeaderParams,
-    tx: &mut PgTransaction<'_>,
-    _srv: &LTZFServer,
-) -> Result<KalGetResponse> {
-    // input validation
-    let result = find_applicable_date_range(
-        qparams.y.map(|x| x as u32),
-        qparams.m.map(|x| x as u32),
-        qparams.dom.map(|x| x as u32),
-        qparams.since,
-        qparams.until,
-        hparams.if_modified_since,
-    );
-    if result.is_none() {
-        return Ok(KalGetResponse::Status416_RequestRangeNotSatisfiable {
-            x_rate_limit_limit: None,
-            x_rate_limit_remaining: None,
-            x_rate_limit_reset: None,
-        });
-    }
-
-    let params = retrieve::SitzungFilterParameters {
-        gremium_like: qparams.gr.clone(),
-        limit: qparams.limit.map(|x| x as u32),
-        offset: qparams.offset.map(|x| x as u32),
-        parlament: qparams.p,
-        vgid: None,
-        wp: qparams.wp.map(|x| x as u32),
-        since: result.as_ref().unwrap().since,
-        until: result.unwrap().until,
-    };
-
-    // retrieval
-    let result = retrieve::sitzung_by_param(&params, tx).await?;
-    if result.is_empty() {
-        Ok(KalGetResponse::Status204_NoContent {
-            x_rate_limit_limit: None,
-            x_rate_limit_remaining: None,
-            x_rate_limit_reset: None,
-        })
-    } else {
-        Ok(KalGetResponse::Status200_SuccessfulResponse {
-            body: result,
-            x_rate_limit_limit: None,
-            x_rate_limit_remaining: None,
-            x_rate_limit_reset: None,
-            x_total_count: (),
-            x_total_pages: (),
-            x_page: (),
-            x_per_page: (),
-            link: (),
-        })
-    }
-}
-
-#[cfg(test)]
-mod test_applicable_date_range {
-    use super::find_applicable_date_range;
-    use chrono::DateTime;
-
-    #[test]
-    fn test_date_range_none() {
-        let result = find_applicable_date_range(None, None, None, None, None, None);
-        assert!(
-            result.is_some()
-                && result.as_ref().unwrap().since.is_none()
-                && result.unwrap().until.is_none(),
-            "None dates should not fail but produce (None, None)"
-        );
-    }
-    #[test]
-    fn test_date_range_untilsince() {
-        let since = DateTime::parse_from_rfc3339("1960-01-01T00:00:00+00:00")
-            .unwrap()
-            .to_utc();
-        let until = DateTime::parse_from_rfc3339("1960-01-02T00:00:00+00:00")
-            .unwrap()
-            .to_utc();
-        let result = find_applicable_date_range(None, None, None, Some(since), Some(until), None);
-        assert!(
-            result.is_some()
-                && result.as_ref().unwrap().since == Some(since)
-                && result.unwrap().until == Some(until),
-            "Since and until should yield (since, until)"
-        )
-    }
-    #[test]
-    fn test_date_range_ymd() {
-        let y = 2012u32;
-        let m = 5u32;
-        let d = 12u32;
-
-        // ymd
-        let result = find_applicable_date_range(
-            Some(y as u32),
-            Some(m as u32),
-            Some(d as u32),
-            None,
-            None,
-            None,
-        );
-        let expected_since = chrono::NaiveDate::from_ymd_opt(y as i32, m, d)
-            .unwrap()
-            .and_hms_opt(0, 0, 0)
-            .unwrap()
-            .and_utc();
-        let expected_until = chrono::NaiveDate::from_ymd_opt(y as i32, m, d)
-            .unwrap()
-            .and_hms_opt(23, 59, 59)
-            .unwrap()
-            .and_utc();
-        assert!(result.is_some());
-        let result = result.unwrap();
-        assert!(
-            result.since == Some(expected_since) && result.until == Some(expected_until),
-            "ymd should start and end at the date range"
-        );
-        // ym
-        let result =
-            find_applicable_date_range(Some(y as u32), Some(m as u32), None, None, None, None);
-        let expected_since = chrono::NaiveDate::from_ymd_opt(y as i32, m, 1)
-            .unwrap()
-            .and_hms_opt(0, 0, 0)
-            .unwrap()
-            .and_utc();
-        let expected_until = chrono::NaiveDate::from_ymd_opt(y as i32, m, 31)
-            .unwrap()
-            .and_hms_opt(23, 59, 59)
-            .unwrap()
-            .and_utc();
-        assert!(result.is_some());
-        let result = result.unwrap();
-        assert!(
-            result.since == Some(expected_since) && result.until == Some(expected_until),
-            "ymd should start and end at the date range"
-        );
-        // y
-        let result = find_applicable_date_range(Some(y as u32), None, None, None, None, None);
-        let expected_since = chrono::NaiveDate::from_ymd_opt(y as i32, 1, 1)
-            .unwrap()
-            .and_hms_opt(0, 0, 0)
-            .unwrap()
-            .and_utc();
-        let expected_until = chrono::NaiveDate::from_ymd_opt(y as i32, 12, 31)
-            .unwrap()
-            .and_hms_opt(23, 59, 59)
-            .unwrap()
-            .and_utc();
-        assert!(result.is_some());
-        let result = result.unwrap();
-        assert!(
-            result.since == Some(expected_since) && result.until == Some(expected_until),
-            "ymd should start and end at the date range"
-        );
-    }
-
-    #[test]
-    fn test_minmax() {
-        let y = 2012u32;
-
-        let since = chrono::NaiveDate::from_ymd_opt(2000, 3, 1)
-            .unwrap()
-            .and_hms_opt(0, 0, 0)
-            .unwrap()
-            .and_utc();
-        let until = chrono::NaiveDate::from_ymd_opt(2012, 7, 31)
-            .unwrap()
-            .and_hms_opt(15, 59, 59)
-            .unwrap()
-            .and_utc();
-
-        let expected_since = chrono::NaiveDate::from_ymd_opt(2012, 1, 1)
-            .unwrap()
-            .and_hms_opt(0, 0, 0)
-            .unwrap()
-            .and_utc();
-        let expected_until = chrono::NaiveDate::from_ymd_opt(2012, 7, 31)
-            .unwrap()
-            .and_hms_opt(15, 59, 59)
-            .unwrap()
-            .and_utc();
-
-        let result =
-            find_applicable_date_range(Some(y), None, None, Some(since), Some(until), None);
-        assert!(result.is_some());
-        let result = result.unwrap();
-        assert!(result.since.is_some() && result.since.unwrap() == expected_since);
-        assert!(result.until.is_some() && result.until.unwrap() == expected_until);
     }
 }
 
@@ -978,12 +676,12 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::KalGetQueryParams {
+                        page: None,
+                        per_page: None,
                         y: Some(recent_date.format("%Y").to_string().parse::<i32>().unwrap()),
                         m: Some(recent_date.format("%m").to_string().parse::<i32>().unwrap()),
                         dom: None,
                         gr: None,
-                        limit: Some(10),
-                        offset: Some(0),
                         p: Some(models::Parlament::Bt),
                         since: None,
                         until: None,
@@ -1014,12 +712,12 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::KalGetQueryParams {
+                        page: None,
+                        per_page: None,
                         y: None,
                         m: None,
                         dom: None,
                         gr: None,
-                        limit: None,
-                        offset: None,
                         p: None,
                         since: Some(chrono::Utc::now()),
                         until: Some(chrono::Utc::now() - chrono::Duration::days(1)), // until is before since
@@ -1058,12 +756,12 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::KalGetQueryParams {
+                        page: None,
+                        per_page: None,
                         y: None,
                         m: None,
                         dom: None,
                         gr: None,
-                        limit: None,
-                        offset: None,
                         p: Some(models::Parlament::Bt),
                         since: Some(start_date),
                         until: Some(end_date),
@@ -1236,8 +934,8 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::SGetQueryParams {
-                        limit: None,
-                        offset: None,
+                        page: None,
+                        per_page: None,
                         p: Some(models::Parlament::Bt),
                         since: None,
                         until: None,
@@ -1267,8 +965,8 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::SGetQueryParams {
-                        limit: None,
-                        offset: None,
+                        page: None,
+                        per_page: None,
                         p: None,
                         since: Some(Utc::now()),
                         until: Some(Utc::now() - chrono::Duration::days(365)),
@@ -1324,8 +1022,8 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::SGetQueryParams {
-                        limit: Some(10),
-                        offset: Some(0),
+                        page: None,
+                        per_page: None,
                         p: Some(models::Parlament::Bt),
                         since: None,
                         until: None,
@@ -1353,8 +1051,8 @@ mod sitzung_test {
                         if_modified_since: Some(chrono::Utc::now()),
                     },
                     &models::SGetQueryParams {
-                        limit: Some(10),
-                        offset: Some(0),
+                        page: None,
+                        per_page: None,
                         p: Some(models::Parlament::Bt),
                         since: None,
                         until: None,
@@ -1384,8 +1082,8 @@ mod sitzung_test {
                         if_modified_since: None,
                     },
                     &models::SGetQueryParams {
-                        limit: Some(10),
-                        offset: Some(0),
+                        page: None,
+                        per_page: None,
                         p: Some(models::Parlament::Bt),
                         since: None,
                         until: None,
@@ -1569,7 +1267,11 @@ mod sitzung_test {
                 .unwrap();
             assert_eq!(
                 response,
-                SitzungDeleteResponse::Status404_NoElementWithThisID
+                SitzungDeleteResponse::Status404_NotFound {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
             );
 
             // - Delete session with insufficient permissions
